code is fixed for the Cart Page
