+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "jestest": "jest",
+    "copy": "Xcopy /E /I static public/static",
     "dev": "webpack serve --hot",
     "build": "rimraf ./src/dist && webpack --mode production",
     "server": "canned -p 5000 ./server",
